requirements.txt
anthropic
openai
requests
flask
python-telegram-bot
python-dotenv
gunicorn
